split up hashing and remove gethashes
