+++ content/strings/HashInterval.h
@@ -1,0 +1,22 @@
+/**
+ * Author: Simon Lindholm
+ * Date: 2015-03-15
+ * License: CC0
+ * Source: own work
+ * Description: Various self-explanatory methods for string hashing.
+ * Status: stress-tested
+ */
+#include "Hashing.h"
+
+struct HashInterval {
+	vector<H> ha, pw;
+	HashInterval(string& str) : ha(sz(str)+1), pw(ha) {
+		pw[0] = 1;
+		rep(i,0,sz(str))
+			ha[i+1] = ha[i] * C + str[i],
+			pw[i+1] = pw[i] * C;
+	}
+	H hashInterval(int a, int b) { // hash [a, b)
+		return ha[b] - ha[a] * pw[b - a];
+	}
+};
