+++ content/strings/chapter.tex
@@ -9,5 +9,7 @@
 \kactlimport{SuffixAutomaton.h}
 \kactlimport{SuffixTree.h}
 \kactlimport{Hashing.h}
+\kactlimport{HashInterval.h}
+% \kactlimport{getHashes.h}
 \kactlimport{AhoCorasick.h}
 \kactlimport{AhoCorasick-Tyler.h}
