+++ content/strings/getHashes.h
@@ -1,0 +1,21 @@
+/**
+ * Author: Simon Lindholm
+ * Date: 2015-03-15
+ * License: CC0
+ * Source: own work
+ * Description: Various self-explanatory methods for string hashing.
+ * Status: stress-tested
+ */
+#include <Hashing.h>
+
+vector<H> getHashes(string& str, int length) {
+	if (sz(str) < length) return {};
+	H h = 0, pw = 1;
+	rep(i,0,length)
+		h = h * C + str[i], pw = pw * C;
+	vector<H> ret = {h};
+	rep(i,length,sz(str)) {
+		ret.push_back(h = h * C + str[i] - pw * str[i-length]);
+	}
+	return ret;
+}
